+++ public/index.html
@@ -28,7 +28,9 @@
             <b>R</b>: R-key click, resume animation <br>
             <b>P</b>: P-key click, pause animation <br>
             <b>Q</b>: Q-key click, reset scene, clear animation <br>
+            
             <b>Space</b>: Space bar click, add frame <br>
+            <b>E</b>: E-key click, delete frame <br>
 
             <br>
             <b>O</b>: O-key press, rotate objects left<br>
